Time: 1 ms (42.39%), Space: 15.3 MB (91.08%) - LeetHub

+++ 2277-count-equal-and-divisible-pairs-in-an-array/2277-count-equal-and-divisible-pairs-in-an-array.cpp
@@ -5,7 +5,7 @@ public:
         int c = 0;
         for (int i = 0; i < n; i++) {
             for (int j = i + 1; j < n; j++) {
-                if (nums[i] == nums[j] && (long long)(i * j) % k == 0) {
+                if (nums[i] == nums[j] && (i * j) % k == 0) {
                     c++;
                 }
             }
